refactor: fixed typos & eslint fixes
