+++ src/api/hooks.ts
@@ -1,6 +1,6 @@
-import { CustomFeatures, CustomGuildInfo } from '../config/types';
+import { CustomFeatures, CustomGuildInfo } from '@/config/types';
 import { QueryClient, useMutation, useQuery } from '@tanstack/react-query';
-import { getGuild, getGuilds, fetchUserInfo } from '@/api/discord';
+import { getGuilds, fetchUserInfo } from '@/api/discord';
 import {
 	disableFeature,
 	enableFeature,
@@ -31,21 +31,9 @@ export const Keys = {
 	login: ['login'],
 	guild_info: (guild: string) => ['guild_info', guild],
 	features: (guild: string, feature: string) => ['feature', guild, feature],
-	guildRoles: (guild: string) => ['gulid_roles', guild],
-	guildChannels: (guild: string) => ['gulid_channel', guild],
+	guildRoles: (guild: string) => ['guild_roles', guild],
+	guildChannels: (guild: string) => ['guild_channel', guild],
 };
-
-export const Mutations = {
-	updateFeature: (guild: string, id: string) => ['feature', guild, id],
-};
-
-export function useGuild(id: string) {
-	const accessToken = useAccessToken();
-
-	return useQuery(['guild', id], () => getGuild(accessToken as string, id), {
-		enabled: accessToken != null,
-	});
-}
 
 export function useGuilds() {
 	const accessToken = useAccessToken();
@@ -100,20 +88,20 @@ export function useEnableFeatureMutation() {
 			async onSuccess(_, { guild, feature, enabled }) {
 				await client.invalidateQueries(Keys.features(guild, feature));
 				client.setQueryData<GuildInfo | null>(Keys.guild_info(guild), (prev) => {
-					if (prev == null) return null;
+					if (prev === null) return null;
 
 					if (enabled) {
 						return {
 							...prev,
-							enabledFeatures: prev.enabledFeatures.includes(feature)
+							enabledFeatures: prev?.enabledFeatures?.includes(feature)
 								? prev.enabledFeatures
-								: [...prev.enabledFeatures, feature],
+								: [...prev?.enabledFeatures as string[], feature],
 						};
 					}
 					else {
 						return {
 							...prev,
-							enabledFeatures: prev.enabledFeatures.filter((f) => f !== feature),
+							enabledFeatures: prev?.enabledFeatures?.filter((f) => f !== feature),
 						};
 					}
 				});
